bug fixing migration to 26 SDK update build tools to 27.0.3 update support library to 27.0.2

+++ rendererrecyclerviewadapter/src/main/java/com/github/vivchar/rendererrecyclerviewadapter/binder/ViewFinder.java
@@ -105,7 +105,7 @@ public interface ViewFinder {
 	@NonNull
 	ViewFinder setVisibility(@IdRes int ID, int visibility);
 	@NonNull
-	ViewFinder setAlpha(@IdRes int ID, int alpha);
+	ViewFinder setAlpha(@IdRes int ID, float alpha);
 	@NonNull
 	ViewFinder addView(@IdRes int ID, View child);
 	@NonNull
